Fixed 'cycles'.
We actually shouldn't be modifying the stack at all, just recurring with
a pushed argument and popping it when finished. Turns out this works out
much nicer if we have the stack be from a Reader instead. RWS simplified
the code considerably.

This is the first time I ever used RWS in uncontrived code! cool!